experimental-pairing-gen: rename files, clean up directory structure of template generators
