+++ internal/generators/curve/generated-code.go
