+++ internal/generators/gpoint/templates/add.go
