+++ internal/generators/gpoint/templates/base.go
