+++ internal/generators/gpoint/templates/benchmarks.go
