+++ internal/generators/gpoint/templates/double.go
