+++ internal/generators/gpoint/templates/scalarmul.go
