+++ internal/generators/gpoint/templates/tests.go
