+++ internal/generators/pairing/generated-code.go
