+++ internal/generators/pairing/templates/expt.go
