+++ internal/generators/pairing/templates/frobenius.go
