+++ internal/generators/pairing/templates/pairing.go
