+++ internal/generators/pairing/templates/tests.go
