+++ internal/generators/tower/generated-code.go
@@ -18,10 +18,4 @@ package tower
 
 import "github.com/consensys/gurvy/bw6_761/fp"
 
-// InitTwoInv set z.TwoInv to the inverse of 2 as an fp.Element
-func (z *Data) InitTwoInv() *Data {
-	var twoInv fp.Element
-	twoInv.SetUint64(2).Inverse(&twoInv)
-	z.TwoInv = twoInv[:]
-	return z
-}
+type fpElement = fp.Element
