+++ internal/generators/tower/templates/fp12/base.go
