+++ internal/generators/tower/templates/fp12/inline.go
