+++ internal/generators/tower/templates/fp2/base.go
