+++ internal/generators/tower/templates/fp2/inline.go
