+++ internal/generators/tower/templates/fp6/base.go
