+++ internal/generators/tower/templates/fp6/inline.go
