+++ internal/generators/tower/templates/fp6/mul.go
